core/chore: strongly type public keys
Strongly type public keys for domain models to provide better ergonomics
and ensure that they are semantically valid. This exposes a design flaw
that has to be fixed before implementing the codec module:
Deserialization should only fail for malformed messages but not for
protocol validation as this is the wrong boundary. This means we'll
split the core module protocol types into wire and model (i.e dto and
domain).