core/fix: handle event id serialization edge case
NIP-01 explicitly says to produce a **utf-8 json string** and to **only
escape** certain characters. This implies the non-mentioned control
characters should not be escaped which is in conflict with producing a
valid utf-8 json string. This SDK now decides to throw an exception when
encountering characters that may lead to inconsistent ids across clients
so no such events can be created with this SDK and events comming in
from the wire are considered valid but trying to call something like
`event.verify()` will throw with a good explanation.

Also see: https://github.com/nostr-protocol/nips/issues/354