core: remove template upgraders
Trying to add a new nip-module I noticed that the current setup with the
event template upgraders does not really work in practice and its really
complicated to get it to work. Then I noticed there is actually no
reason why we'd want consumers to allow to create a generic event
template and then try to upgrade it. Sure this is nice flexibility and
maybe there are very rare edge cases where you'd want to do that but
generally, if you want to work with an event template just use the
modules builder methods to get the correct template.