core: add nip module context
The nip module context will provide json codec for content / payload
parsing as well as other stuff that might need to be added later (for
example crypto). This has the benefit that we avoid each nip module
creating their own json or crypto instance. I do not like the ergonomics
of having to pass this context to `MyEventView.template(ctx)` for
example...but cant think of a better solution.